前端优化打包日志问题
